ema-thresholding-ml-ids/requirements.txt
# Core ML and data science libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
scipy>=1.7.0

# Data processing and I/O
PyYAML>=6.0
requests>=2.25.0

# Logging and utilities
tqdm>=4.62.0

# Development and testing
pytest>=6.0.0
pytest-cov>=2.12.0

# Pre-commit hooks
pre-commit>=2.15.0

# Optional: Visualization (for analysis scripts)
matplotlib>=3.4.0
seaborn>=0.11.0

# Optional: Jupyter notebooks
jupyter>=1.0.0
ipywidgets>=7.6.0
